Fix all format signed/unsigned inconsistencies
They are harmless, but fixed for consistency and as a preparation to
a stricter format check.

Type of inconsistencies:
Using %zd for size_t instead of %zu.
Using %d for enum type, which is implementation defined (unsigned int in gcc).
Using %d for unsigned int.
Using %x for int.

+++ link-grammar/error.c
@@ -246,7 +246,7 @@ static const char *error_severity_label(lg_error_severity sev)
 	}
 	else if ((sev < 1) || (sev > lg_None))
 	{
-		snprintf(sevlabel, MAX_SEVERITY_LABEL_SIZE, "Message severity %d", sev);
+		snprintf(sevlabel, MAX_SEVERITY_LABEL_SIZE, "Message severity %d", (int)sev);
 	}
 	else
 	{
